Move annotation.py to h.search.docs & split into 3
Move annotation.py into h.search.docs area. This will contain
elasticsearch documents. annotation.py currently contains 3 docs
so split this into 3 seperate doc files (technically the Target
and Selector are InnerDocs of Annotation but they are still seperate
docs even though they are nested).
